定时检查TCP状态，打印单独TCP LOG

+++ Drill/Protocol/WITSEnc.h
@@ -22,5 +22,5 @@ private:
 	static string EncWITSFixHead(UINT SeqNO, WITSCFG* ptWITS);
 	static string EncOnlyTorqData(WITSCFG* ptWITS, WITSRPTDATA* ptData);
 	static string EncWITSFixTail();
-	static string EncHisFixHead(UINT SeqNO, WITSCFG* ptWITS, TorqData::Torque* ptTorq, bool bBreakout, double diff = 0);
+	static string EncHisFixHead(UINT SeqNO, WITSCFG* ptWITS, TorqData::Torque* ptTorq, bool bBreakout, int diff = 0);
 };
